[FEAT] Refactors Page service to support nested routes
This PR refactors the page service to allow it to support nested routes.
The guiding principle here is that our UX for traversing the guides is
a depth-first traversal of the docs tree, so rather than traversing the
tree directly each time, we do a single traversal once, creating both a
convenient data structure for looking up pages based on URL, and setting
up the next/prev links between each page.

Should likely add some tests around the `nextPage` and `previousPage`
and their sections as well, to ensure that next/prev page work with
arbitrary navigation trees.