[BUGFIX] Fix page.previousSection and page.nextSection
The refactor to the `page` service changed the meaning of
`previousSection` and `nextSection` to mean the section of the next page
and previous page. This would be a pretty major change, and caused some
breakage in the existing themes, so this PR restores them to mean the
actual next/previous section. This works for nested sections, and tests
have been added for them.

In addition, a few new boolean values have been added to the service:

* `previousIsFirstPage'
* `previousIsLastPage'
* `nextIsFirstPage`
* `nextIsLastPage'

These are necessary for themes to be able to update to using nested
sections, since they can no longer customize their next/previous chapter
messages based solely on the current page.

This API is pretty verbose, but I didn't want to expose the page objects
directly without more input from @mansona since it would be a much
larger change.

Also includes more validations of the page structure, and more tests.